+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 ## v1.2.0 2026-05-30
 
 Added:
+- added methods to struns
 - added enums
 - added ptr access to struct fields and passing by reference with &
 - added varargs ... 
-- expanded io stdlib
+- expanded io stdlib to include more file io functions
 
 Removed: